Add files via upload
Solves a custom maths puzzle with the predicate, puzzle_solution/1.
The maths puzzle is defined below:

A maths puzzle is a square grid of squares, each to be filled in with a single digit 1–9 (zero is not permitted) satisfying these constraints:
    -each row and each column contains no repeated digits;
    -all squares on the diagonal line from upper left to lower right contain the same value;
and
    -the heading of reach row and column (leftmost square in a row and topmost square in a column) holds either the sum or the product of all the digits in that row or column
Note that the row and column headings are not considered to be part of the row or column, and so may be filled with a number larger than a single digit. The upper left corner of the puzzle is not meaningful.
When the puzzle is originally posed, most or all of the squares will be empty, with the headings filled in. The goal of the puzzle is to fill in all the squares according to the rules. A proper maths puzzle will have at most one solution.

I declare no authorship of the puzzle, only the implementation of puzzle_solution/1.